%    Copyright (c) 2017 Russell V. Lenth

\name{unrepx-package}
\alias{unrepx-package}
\alias{unrepx}
\docType{package}
\title{
Analysis and graphics for unreplicated experiments
}

\description{
Provides half-normal plots, reference plots, and Pareto plots
    of effects from an unreplicated experiment, along with various PSE
    measures, simulated reference distributions, and other tools
}
\details{
See the package vignette: \code{vignette("overview", package="unrepx")}
}
\author{
Russell V. Lenth

Maintainer: Russell V. Lenth <russell-lenth@uiowa.edu>
}

%%\references{}


\keyword{package }
